Fix topic model and LIWC score functions
The topic model score function broke when there the number of topics that
could be computed for a given text was too small. Now it returns a 0 value
instead for every topic above the actual number it can compute.

The LIWC score function broke for certain inputs that made the REST API return
an error status. Now it returns a zero vector instead. Proper handling for
HTTP status 429, too many requests, was also added.